Add Football Score Prediction Project
model.py is our first model. After that, I modified it to predict 3 classes with result_head using softmax function in model2.py, and then, in model3.py I've made 3 heads for result prediction. app.py is our application. model3.py has to be run to produce a model that can be used to run the app with app.py